+++ contacts_template/lib/main.dart
@@ -1,0 +1,9 @@
+import 'package:contacts_template/src/view.dart';
+
+void main() => runApp(
+      MyApp(),
+      errorHandler: null,
+      errorScreen: null,
+      errorReport: null,
+      allowNewHandlers: false,
+    );