+++ contacts_template/lib/src/app/view.dart
@@ -1,0 +1,2 @@
+// application's main app
+export 'package:contacts_template/src/app/view/app.dart';