+++ contacts_template/lib/src/controller.dart
@@ -1,0 +1,5 @@
+// The MVC application framework.
+export 'package:mvc_application/controller.dart';
+
+// The controller for the home screen.
+export 'package:contacts_template/src/home/controller.dart';