+++ contacts_template/lib/src/home/controller.dart
@@ -1,0 +1,7 @@
+///
+/// The controllers involved for the home screen.
+///
+
+
+// The Contacts app
+export 'package:contacts_template/src/home/controller/contacts_controller.dart';