+++ contacts_template/lib/src/home/model.dart
@@ -1,0 +1,9 @@
+///
+/// All the data involved for the home screen.
+///
+
+export 'package:contacts_template/src/home/model/contacts/contact.dart';
+
+export 'package:contacts_template/src/home/model/contacts/contact_fields.dart';
+
+export 'package:contacts_template/src/home/model/contacts/contacts_db.dart';