+++ contacts_template/lib/src/home/view.dart
@@ -1,0 +1,6 @@
+///
+/// All the UI involved in the home screen.
+///
+
+// Contacts app
+export 'package:contacts_template/src/home/view/contacts_view.dart';