+++ contacts_template/lib/src/model.dart
@@ -1,0 +1,5 @@
+// The MVC application framework.
+export 'package:mvc_application/model.dart';
+
+// The data for the home screen.
+export 'package:contacts_template/src/home/model.dart';