+++ contacts_template/lib/src/view.dart
@@ -1,0 +1,8 @@
+// The mvc application framework.
+export 'package:mvc_application/view.dart';
+
+// The UI at the app level.
+export 'package:contacts_template/src/app/view.dart' hide ISOSpinner;
+
+// The UI for the home screen.
+export 'package:contacts_template/src/home/view.dart';